Remove bashisms and add safeguards

+++ install.sh
@@ -1,16 +1,18 @@
 #!/bin/sh
 
-if [[ $(uname) != 'Darwin' ]]
+if [ "$(uname)" != "Darwin" ]
 then
     DIR=$(dirname $(realpath -s "$0"))
 else
     DIR=$(dirname $(realpath "$0"))
 fi
 
-STUFF="zshrc zsh vimrc vim tmux.conf"
-for i in $STUFF
+for i in zshrc zsh vimrc vim tmux.conf
 do
-    ln -sfn "$DIR/$i" "$HOME/.$i"
+    if ! [ -e "$HOME/.$i" ]
+    then
+        ln -sfn "$DIR/$i" "$HOME/.$i"
+    fi
 done
 
 if command -v htop > /dev/null
@@ -19,9 +21,9 @@ then
     ln -sfn "$DIR/htoprc" "$HOME/.config/htop/htoprc"
 fi
 
-if command -v nvim > /dev/null
+NVIM_CONFIG="$HOME/.config/nvim"
+if command -v nvim > /dev/null && ! [ -e "$NVIM_CONFIG/init.vim" ]
 then
-    NVIM_CONFIG="$HOME/.config/nvim"
     mkdir -p "$NVIM_CONFIG"
     ln -sfn "$DIR/vimrc" "$NVIM_CONFIG/init.vim"
     ln -sfn "$DIR/vim/autoload" "$NVIM_CONFIG/autoload"
